feat(workflows): improve display image modal
When you have of 1 image sometimes when you click in the images to have
a bigger view, you want to move between images with 1 click to compare
them and decide what achieve more your needs.

Currently is not possible because you can only open 1 at the time, so
you need to close the modal and open a different image.